// packages/ocom/graphql/src/schema/builder/generated.ts
export type Maybe<T> = T | null;
// ...
export type Omit<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>;
export type RequireFields<T, K extends keyof T> = Omit<T, K> & { [P in K]-?: NonNullable<T[P]> };
// ...
export type BlobAuthHeader = {
  readonly __typename?: "BlobAuthHeader";
  readonly authHeader?: Maybe<Scalars["String"]["output"]>;
  readonly blobName?: Maybe<Scalars["String"]["output"]>;
  readonly blobPath?: Maybe<Scalars["String"]["output"]>;
  readonly indexTags?: Maybe<ReadonlyArray<Maybe<BlobIndexTag>>>;
  readonly metadataFields?: Maybe<ReadonlyArray<Maybe<BlobMetadataField>>>;
  readonly requestDate?: Maybe<Scalars["String"]["output"]>;
};

export type BlobIndexTag = {
  readonly __typename?: "BlobIndexTag";
  readonly name: Scalars["String"]["output"];
  readonly value: Scalars["String"]["output"];
};

export type BlobMetadataField = {
  readonly __typename?: "BlobMetadataField";
  readonly name: Scalars["String"]["output"];
  readonly value: Scalars["String"]["output"];
};

/**  Required to enable Apollo Cache Control  */
export type CacheControlScope = "PRIVATE" | "PUBLIC";

export type Community = MongoBase & {
  readonly __typename?: "Community";
  readonly createdAt: Scalars["DateTime"]["output"];
  readonly createdBy: EndUser;
  readonly domain?: Maybe<Scalars["String"]["output"]>;
  readonly handle?: Maybe<Scalars["String"]["output"]>;
  readonly id: Scalars["ObjectID"]["output"];
  readonly name: Scalars["String"]["output"];
  readonly publicContentBlobUrl?: Maybe<Scalars["String"]["output"]>;
  readonly schemaVersion: Scalars["String"]["output"];
  readonly updatedAt: Scalars["DateTime"]["output"];
  readonly whiteLabelDomain?: Maybe<Scalars["String"]["output"]>;
};

export type CommunityCreateInput = {
  readonly name: Scalars["String"]["input"];
};

export type CommunityMutationResult = MutationResult & {
  readonly __typename?: "CommunityMutationResult";
  readonly community?: Maybe<Community>;
  readonly status: MutationStatus;
};

export type EndUser = MongoBase & {
  readonly __typename?: "EndUser";
  readonly accessBlocked?: Maybe<Scalars["Boolean"]["output"]>;
  readonly createdAt?: Maybe<Scalars["DateTime"]["output"]>;
  readonly displayName?: Maybe<Scalars["String"]["output"]>;
  readonly externalId?: Maybe<Scalars["String"]["output"]>;
  readonly id: Scalars["ObjectID"]["output"];
  readonly personalInformation?: Maybe<EndUserPersonalInformation>;
  readonly schemaVersion?: Maybe<Scalars["String"]["output"]>;
  readonly tags?: Maybe<ReadonlyArray<Maybe<Scalars["String"]["output"]>>>;
  readonly updatedAt?: Maybe<Scalars["DateTime"]["output"]>;
};

export type EndUserContactInformation = {
  readonly __typename?: "EndUserContactInformation";
  readonly email: Scalars["String"]["output"];
};

export type EndUserIdentityDetails = {
  readonly __typename?: "EndUserIdentityDetails";
  readonly lastName: Scalars["String"]["output"];
  readonly legalNameConsistsOfOneName: Scalars["Boolean"]["output"];
  readonly restOfName?: Maybe<Scalars["String"]["output"]>;
};

export type EndUserPersonalInformation = {
  readonly __typename?: "EndUserPersonalInformation";
  readonly contactInformation?: Maybe<EndUserContactInformation>;
  readonly identityDetails?: Maybe<EndUserIdentityDetails>;
};
// ...
export type MutationResult = {
  readonly status: MutationStatus;
};
// ...
export type ResolversObject<TObject> = WithIndex<TObject>;
// ...
export type BlobAuthHeaderResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["BlobAuthHeader"] = ResolversParentTypes["BlobAuthHeader"],
> = ResolversObject<{
  authHeader?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  blobName?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  blobPath?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  indexTags?: Resolver<Maybe<ReadonlyArray<Maybe<ResolversTypes["BlobIndexTag"]>>>, ParentType, ContextType>;
  metadataFields?: Resolver<Maybe<ReadonlyArray<Maybe<ResolversTypes["BlobMetadataField"]>>>, ParentType, ContextType>;
  requestDate?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type BlobIndexTagResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["BlobIndexTag"] = ResolversParentTypes["BlobIndexTag"],
> = ResolversObject<{
  name?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  value?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type BlobMetadataFieldResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["BlobMetadataField"] = ResolversParentTypes["BlobMetadataField"],
> = ResolversObject<{
  name?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  value?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export interface ByteScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes["Byte"], any> {
  name: "Byte";
}

export type CommunityResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["Community"] = ResolversParentTypes["Community"],
> = ResolversObject<{
  createdAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  createdBy?: Resolver<ResolversTypes["EndUser"], ParentType, ContextType>;
  domain?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  handle?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  id?: Resolver<ResolversTypes["ObjectID"], ParentType, ContextType>;
  name?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  publicContentBlobUrl?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  schemaVersion?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  updatedAt?: Resolver<ResolversTypes["DateTime"], ParentType, ContextType>;
  whiteLabelDomain?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type CommunityMutationResultResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["CommunityMutationResult"] = ResolversParentTypes["CommunityMutationResult"],
> = ResolversObject<{
  community?: Resolver<Maybe<ResolversTypes["Community"]>, ParentType, ContextType>;
  status?: Resolver<ResolversTypes["MutationStatus"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
// ...
export type EndUserResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["EndUser"] = ResolversParentTypes["EndUser"],
> = ResolversObject<{
  accessBlocked?: Resolver<Maybe<ResolversTypes["Boolean"]>, ParentType, ContextType>;
  createdAt?: Resolver<Maybe<ResolversTypes["DateTime"]>, ParentType, ContextType>;
  displayName?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  externalId?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  id?: Resolver<ResolversTypes["ObjectID"], ParentType, ContextType>;
  personalInformation?: Resolver<Maybe<ResolversTypes["EndUserPersonalInformation"]>, ParentType, ContextType>;
  schemaVersion?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  tags?: Resolver<Maybe<ReadonlyArray<Maybe<ResolversTypes["String"]>>>, ParentType, ContextType>;
  updatedAt?: Resolver<Maybe<ResolversTypes["DateTime"]>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type EndUserContactInformationResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["EndUserContactInformation"] = ResolversParentTypes["EndUserContactInformation"],
> = ResolversObject<{
  email?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type EndUserIdentityDetailsResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["EndUserIdentityDetails"] = ResolversParentTypes["EndUserIdentityDetails"],
> = ResolversObject<{
  lastName?: Resolver<ResolversTypes["String"], ParentType, ContextType>;
  legalNameConsistsOfOneName?: Resolver<ResolversTypes["Boolean"], ParentType, ContextType>;
  restOfName?: Resolver<Maybe<ResolversTypes["String"]>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;

export type EndUserPersonalInformationResolvers<
  ContextType = GraphContext,
  ParentType extends ResolversParentTypes["EndUserPersonalInformation"] = ResolversParentTypes["EndUserPersonalInformation"],
> = ResolversObject<{
  contactInformation?: Resolver<Maybe<ResolversTypes["EndUserContactInformation"]>, ParentType, ContextType>;
  identityDetails?: Resolver<Maybe<ResolversTypes["EndUserIdentityDetails"]>, ParentType, ContextType>;
  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
}>;
// ...
export type Resolvers<ContextType = GraphContext> = ResolversObject<{
  AccountNumber?: GraphQLScalarType;
  BigInt?: GraphQLScalarType;
  BlobAuthHeader?: BlobAuthHeaderResolvers<ContextType>;
  BlobIndexTag?: BlobIndexTagResolvers<ContextType>;
  BlobMetadataField?: BlobMetadataFieldResolvers<ContextType>;
  Byte?: GraphQLScalarType;
  Community?: CommunityResolvers<ContextType>;
  CommunityMutationResult?: CommunityMutationResultResolvers<ContextType>;
  CountryCode?: GraphQLScalarType;
  CountryName?: GraphQLScalarType;
  Cuid?: GraphQLScalarType;
  Currency?: GraphQLScalarType;
  DID?: GraphQLScalarType;
  Date?: GraphQLScalarType;
  DateTime?: GraphQLScalarType;
  DateTimeISO?: GraphQLScalarType;
  DeweyDecimal?: GraphQLScalarType;
  Duration?: GraphQLScalarType;
  EmailAddress?: GraphQLScalarType;
  EndUser?: EndUserResolvers<ContextType>;
  EndUserContactInformation?: EndUserContactInformationResolvers<ContextType>;
  EndUserIdentityDetails?: EndUserIdentityDetailsResolvers<ContextType>;
  EndUserPersonalInformation?: EndUserPersonalInformationResolvers<ContextType>;
  GUID?: GraphQLScalarType;
  GeoJSON?: GraphQLScalarType;
  HSL?: GraphQLScalarType;
  HSLA?: GraphQLScalarType;
  HexColorCode?: GraphQLScalarType;
  Hexadecimal?: GraphQLScalarType;
  IBAN?: GraphQLScalarType;
  IP?: GraphQLScalarType;
  IPCPatent?: GraphQLScalarType;
  IPv4?: GraphQLScalarType;
  IPv6?: GraphQLScalarType;
  ISBN?: GraphQLScalarType;
  ISO8601Duration?: GraphQLScalarType;
  JSON?: GraphQLScalarType;
  JSONObject?: GraphQLScalarType;
  JWT?: GraphQLScalarType;
  LCCSubclass?: GraphQLScalarType;
  Latitude?: GraphQLScalarType;
  LocalDate?: GraphQLScalarType;
  LocalDateTime?: GraphQLScalarType;
  LocalEndTime?: GraphQLScalarType;
  LocalTime?: GraphQLScalarType;
  Locale?: GraphQLScalarType;
  Long?: GraphQLScalarType;
  Longitude?: GraphQLScalarType;
  MAC?: GraphQLScalarType;
  Member?: MemberResolvers<ContextType>;
  MemberAccount?: MemberAccountResolvers<ContextType>;
  MemberProfile?: MemberProfileResolvers<ContextType>;
  MongoBase?: MongoBaseResolvers<ContextType>;
  MongoSubdocument?: MongoSubdocumentResolvers<ContextType>;
  Mutation?: MutationResolvers<ContextType>;
  MutationResult?: MutationResultResolvers<ContextType>;
  MutationStatus?: MutationStatusResolvers<ContextType>;
  NegativeFloat?: GraphQLScalarType;
  NegativeInt?: GraphQLScalarType;
  NonEmptyString?: GraphQLScalarType;
  NonNegativeFloat?: GraphQLScalarType;
  NonNegativeInt?: GraphQLScalarType;
  NonPositiveFloat?: GraphQLScalarType;
  NonPositiveInt?: GraphQLScalarType;
  ObjectID?: GraphQLScalarType;
  PhoneNumber?: GraphQLScalarType;
  Port?: GraphQLScalarType;
  PositiveFloat?: GraphQLScalarType;
  PositiveInt?: GraphQLScalarType;
  PostalCode?: GraphQLScalarType;
  Query?: QueryResolvers<ContextType>;
  RGB?: GraphQLScalarType;
  RGBA?: GraphQLScalarType;
  RoutingNumber?: GraphQLScalarType;
  SESSN?: GraphQLScalarType;
  SafeInt?: GraphQLScalarType;
  SemVer?: GraphQLScalarType;
  Time?: GraphQLScalarType;
  TimeZone?: GraphQLScalarType;
  Timestamp?: GraphQLScalarType;
  URL?: GraphQLScalarType;
  USCurrency?: GraphQLScalarType;
  UUID?: GraphQLScalarType;
  UnsignedFloat?: GraphQLScalarType;
  UnsignedInt?: GraphQLScalarType;
  UtcOffset?: GraphQLScalarType;
  Void?: GraphQLScalarType;
}>;